Restructuring gameta's repository, consolidating all non-CLI code into base package
