+++ gameta/base/__init__.py
@@ -1,0 +1,9 @@
+
+from .context import GametaContext, gameta_context
+
+
+__all__ = [
+
+    # Context
+    'GametaContext', 'gameta_context'
+]
