+++ gameta/cli.py
@@ -4,7 +4,7 @@ from os.path import abspath
 
 import click
 
-from .context import gameta_context, GametaContext
+from gameta.base import gameta_context, GametaContext
 from gameta import __version__
 
 
@@ -45,11 +45,3 @@ def gameta_cli(context: GametaContext, project_dir: str, version: bool) -> None:
     context.load()
 
 
-from .init import *
-from .repos import *
-from .tags import *
-from .apply import *
-from .params import *
-from .cmd import *
-from .constants import *
-from .execute import *
