+++ gameta/init.py
@@ -6,7 +6,7 @@ import click
 from git import Repo, InvalidGitRepositoryError, GitError
 
 from .cli import gameta_cli
-from .context import gameta_context, GametaContext
+from gameta.base import gameta_context, GametaContext
 
 
 __all__ = ['init', 'sync']
