+++ gameta/tags.py
@@ -3,7 +3,7 @@ from typing import List, Dict, Tuple
 import click
 
 from .cli import gameta_cli
-from .context import gameta_context, GametaContext
+from gameta.base import gameta_context, GametaContext
 
 
 __all__ = ['tags_cli']
