+++ tests/test_repos.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 
 from click.testing import CliRunner
 
-from gameta.context import GametaContext
+from gameta.base import GametaContext
 from gameta.repos import add, delete, ls, update
